requirements.txt.txt
plotly
requests
tqdm
python-dotenv
